[#13] Minor source code improvements (Part 43: Improved composition of list of files to export)

+++ Main/src/Image.h
@@ -160,7 +160,7 @@
 		BY_SIDES		=2
 	};
 
-	typedef WORD TStdWinError; // Windows standard i/o error
+	typedef DWORD TStdWinError; // Windows standard i/o error
